requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pandas>=2.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.9
